Extract WrappedFd
